feat and test : updated in multiple files and tests
Updates in multiple files to enhance code performance
Updated Test package in an attempt to cover max code

+++ 07-spring-boot-crud-SpringDataJPA-demo/src/main/java/com/learning/spring_boot_demo/crud/service/EmployeeServiceImpl.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 
 import com.learning.spring_boot_demo.crud.dao.EmployeeRepository;
 import com.learning.spring_boot_demo.crud.entity.Employee;
+import com.learning.spring_boot_demo.crud.exception.EmployeeWithMailAlreadyExistsException;
 import com.learning.spring_boot_demo.crud.exception.EmployeeNotFoundException;
 
 @Service
@@ -23,26 +24,42 @@ public class EmployeeServiceImpl implements EmployeeService {
 	//No need to explicitly mention @Transaction as Spring Data JPA
 	//provides this functionality:
 //	@Transactional
-	public List<Employee> findAll() {
+	public List<Employee> findAllEmployees() {
 		return empRepo.findAll();
 	}
 
 	@Override
 //	@Transactional
-	public Employee findById(int id) {
+	public Employee findEmployeeById(int id) {
 		return empRepo.findById(id).
 				orElseThrow(() -> new EmployeeNotFoundException("Employee with Id: "+id+" not found!"));
 	}
 
 	@Override
 //	@Transactional
-	public void save(Employee theEmployee) {
+	public void saveEmployee(Employee theEmployee) {
+		if (empRepo.selectExistsEmail(theEmployee.getEmail()))
+			throw new EmployeeWithMailAlreadyExistsException("Employee with Email: "+theEmployee.getEmail()+" already exists!");
+		
 		empRepo.save(theEmployee);
 	}
 
 	@Override
 //	@Transactional
-	public void deleteById(int id) {
+	public void updateEmployee(Employee theEmployee) {
+		if (!empRepo.existsById(theEmployee.getId()))
+			throw new EmployeeNotFoundException("Employee with Id: "+theEmployee.getId()+" not found!");
+		
+		empRepo.save(theEmployee);
+//		empRepo.setEmployeeInfoById(theEmployee.getFirstName(), theEmployee.getLastName(), theEmployee.getEmail(), id);
+	}
+
+	@Override
+//	@Transactional
+	public void deleteEmployeeById(int id) {
+		if (!empRepo.existsById(id))
+			throw new EmployeeNotFoundException("Employee with Id: "+id+" not found!");
+		
 		empRepo.deleteById(id);
 	}
